RavenDB-18007 Get all shards results
